feat(iv3): retire the IV3 cost report to Shillinq, keep the classification
Implements ADR-081 decision 2. Procest's quarterly IV3 report is deleted:
Iv3ReportService, Iv3ReportController::report(), its route,
Iv3ReportDashboard.vue, the manifest page and menu entry, the
customComponents and menu-layout registrations, both report test suites,
and the `case.kosten` field.

The report could not report anything. Re-verified on the dev instance
before deleting: 18 cases scanned, ZERO carrying a kosten entry, and the
same across all eight quarters of 2025-2026 when this was first found.
Nothing wrote `handling_cost` or `leges_income` at all — the only writer
of `kosten` was VaststellingService's subsidy disbursement, so two of the
report's three columns were structurally incapable of being non-zero.

WHAT IS DELIBERATELY KEPT, and why the ADR's "delete in full" is not
followed literally here. `caseType.iv3Taakveld` is the CLASSIFICATION the
same decision tells procest to keep, and the settings picker that
populates it reads `/api/reports/iv3/taakvelden`. So Iv3TaakveldList and
its JSON stay, and the controller is reduced to that one endpoint
(renamed Iv3TaakveldController; the URL is unchanged because the picker
calls it directly). Deleting the list now would break the field the ADR
preserves. Decision 1 makes Shillinq's BbvTaakveld catalogue the single
authority, and that endpoint is the seam that will read from it — but
Shillinq does not expose it cross-app yet. Sequencing, not an exception.

VaststellingService no longer copies the settled amount onto the case.
That denormalisation existed solely to feed this report. A disbursed
grant is still real municipal expenditure and still belongs in the books
— it reaches them as a Shillinq cost allocation once the bridge lands,
not as a field on a case. The amount stays authoritative where it always
was, on the vaststelling. Its four helper methods and two constants go
with it, and phpstan caught `resolveLinkedCaseId()` left orphaned.

1669 tests green; phpcs, phpstan and psalm clean.
