+++ lib/Service/Subsidie/VaststellingService.php
@@ -8,10 +8,13 @@
  * amount, the accountantsverklaring requirement check, final-bedrag
  * calculation, overpayment detection, and the automatic terugvordering
  * trigger (REQ-SUB-005). The math is pure and unit-tested; finalisation
- * delegates clawback-case creation to TerugvorderingService and,
- * best-effort, auto-populates the linked case's `kosten` with the settled
- * amount (subsidie-settlement-case-costs) via the same ObjectService write
- * path, walking subsidieUitvoering -> subsidieAanvraag -> case.
+ * delegates clawback-case creation to TerugvorderingService.
+ *
+ * It no longer copies the settled amount onto the linked case's `kosten`
+ * array. That denormalisation existed to feed procest's own IV3 report, and
+ * both are gone under ADR-081 — a domain app MUST NOT hold a ledger-shaped
+ * array, and Shillinq is the fleet's only general ledger. The amount stays
+ * authoritative where it always was, on the vaststelling itself.
  *
  * @category Service
  * @package  OCA\Procest\Service\Subsidie
@@ -47,22 +50,6 @@ use Throwable;
  */
 class VaststellingService
 {
-    /**
-     * `case.kosten` entry `type` for a subsidie vaststelling settlement
-     * amount (subsidie-settlement-case-costs). English snake_case to match
-     * the existing `leges_income`/`handling_cost` type-discriminator
-     * convention in {@see \OCA\Procest\Service\Iv3ReportService}; counted
-     * toward `totalCosts` there.
-     */
-    public const KOSTEN_TYPE_SUBSIDY_DISBURSEMENT = 'subsidy_disbursement';
-
-    /**
-     * `case.kosten` entry `source` marker identifying an entry appended by
-     * this service — paired with `vaststellingId` for idempotency (a
-     * re-finalize of the same vaststelling must not duplicate the entry).
-     */
-    public const KOSTEN_SOURCE = 'subsidie_vaststelling';
-
     /**
      * Constructor.
      *
@@ -201,209 +188,21 @@ class VaststellingService
             $clawback = $this->terugvordering->createClawbackCase(uitvoeringId: $uitvoeringId, bedrag: $overpayment);
         }
 
-        // Subsidie-settlement-case-costs: auto-populate the linked case's
-        // kosten with the settled amount. Best-effort — never fails the
-        // vaststelling itself (no linked case, missing chain link, or a
-        // transport error all degrade to a logged warning).
-        $this->appendKostenToLinkedCase(uitvoeringId: $uitvoeringId, vaststellingId: $vaststellingId, bedrag: $vastgesteld);
-
+        // The settled amount used to be appended to the linked case's `kosten`
+        // array, which fed procest's own IV3 report. Both are gone under
+        // ADR-081: a domain app MUST NOT hold a ledger-shaped array, and
+        // Shillinq is the only general ledger. A disbursed grant is real
+        // municipal expenditure and still belongs in the books — it reaches
+        // them as a Shillinq cost allocation, not as a field on a case.
+        // Until that dispatch exists the amount is recorded on the
+        // vaststelling itself (`vastgesteldBedrag`, saved above), which is
+        // where it was always authoritative; the `kosten` copy was a
+        // denormalisation for a report that no longer exists.
         return [
             'vaststelling'   => $saved,
             'terugvordering' => $clawback,
         ];
     }//end finalize()
-
-    /**
-     * Append the settled amount to the case linked via
-     * subsidieUitvoering -> subsidieAanvraag -> case, through the SAME
-     * `ObjectService::saveObject()` write path every other case field
-     * mutation in this app uses. Fail-soft on every branch: an unresolvable
-     * link (no execution id, execution/application/case not found, no case
-     * linked to the application) or any `Throwable` is logged and skipped —
-     * this is enrichment, not a hard dependency of settling.
-     *
-     * Idempotent: a re-finalize of the same vaststelling is detected via
-     * the `source` + `vaststellingId` markers already present on an
-     * existing entry and skipped, so the amount is never duplicated.
-     *
-     * @param string $uitvoeringId   The execution id (may be empty).
-     * @param string $vaststellingId The settlement id (idempotency key).
-     * @param float  $bedrag         The settled amount to record.
-     *
-     * @return void
-     */
-    private function appendKostenToLinkedCase(string $uitvoeringId, string $vaststellingId, float $bedrag): void
-    {
-        if ($uitvoeringId === '' || $bedrag <= 0.0) {
-            return;
-        }
-
-        try {
-            $context = $this->resolveKostenContext();
-            if ($context === null) {
-                return;
-            }
-
-            [$objectService, $register, $uitvoeringSchema, $aanvraagSchema, $caseSchema] = $context;
-
-            $caseId = $this->resolveLinkedCaseId(
-                objectService: $objectService,
-                register: $register,
-                uitvoeringId: $uitvoeringId,
-                uitvoeringSchema: $uitvoeringSchema,
-                aanvraagSchema: $aanvraagSchema
-            );
-            if ($caseId === null) {
-                return;
-            }
-
-            $case = $objectService->find($caseId, register: $register, schema: $caseSchema);
-            if (is_array($case) === false) {
-                return;
-            }
-
-            $kosten = $this->decodeKosten(raw: $case['kosten'] ?? null);
-            if ($this->hasExistingEntry(kosten: $kosten, vaststellingId: $vaststellingId) === true) {
-                // Already recorded for this vaststelling — do not duplicate.
-                return;
-            }
-
-            $kosten[] = [
-                'bedrag'         => round($bedrag, 2),
-                'type'           => self::KOSTEN_TYPE_SUBSIDY_DISBURSEMENT,
-                'datum'          => (new DateTimeImmutable())->format('Y-m-d'),
-                'source'         => self::KOSTEN_SOURCE,
-                'vaststellingId' => $vaststellingId,
-            ];
-
-            $objectService->saveObject(
-                object: ['kosten' => json_encode($kosten)],
-                register: $register,
-                schema: $caseSchema,
-                uuid: $caseId
-            );
-        } catch (Throwable $e) {
-            $this->logger->warning(
-                'Procest subsidie: could not append vaststelling kosten to linked case: '.$e->getMessage()
-            );
-        }//end try
-    }//end appendKostenToLinkedCase()
-
-    /**
-     * Resolve the ObjectService plus every register/schema id the kosten
-     * append needs, or null when any is unavailable/unconfigured (the
-     * fail-soft equivalent of {@see resolve()} — this path must never throw).
-     *
-     * @return array{0: object, 1: string, 2: string, 3: string, 4: string}|null
-     */
-    private function resolveKostenContext(): ?array
-    {
-        $objectService    = $this->settingsService->getObjectService();
-        $register         = $this->settingsService->getConfigValue('register');
-        $uitvoeringSchema = $this->settingsService->getConfigValue('subsidie_uitvoering_schema');
-        $aanvraagSchema   = $this->settingsService->getConfigValue('subsidie_aanvraag_schema');
-        $caseSchema       = $this->settingsService->getConfigValue('case_schema');
-        if ($objectService === null || $register === '' || $uitvoeringSchema === '' || $aanvraagSchema === '' || $caseSchema === '') {
-            return null;
-        }
-
-        return [$objectService, $register, $uitvoeringSchema, $aanvraagSchema, $caseSchema];
-    }//end resolveKostenContext()
-
-    /**
-     * Walk subsidieUitvoering -> subsidieAanvraag -> case to resolve the
-     * linked case id, or null when any hop is missing.
-     *
-     * @param object $objectService    Resolved ObjectService.
-     * @param string $register         Register id.
-     * @param string $uitvoeringId     The execution id.
-     * @param string $uitvoeringSchema The subsidieUitvoering schema id.
-     * @param string $aanvraagSchema   The subsidieAanvraag schema id.
-     *
-     * @return string|null The linked case id, or null when unresolvable.
-     */
-    private function resolveLinkedCaseId(
-        object $objectService,
-        string $register,
-        string $uitvoeringId,
-        string $uitvoeringSchema,
-        string $aanvraagSchema
-    ): ?string {
-        $uitvoering = $objectService->find($uitvoeringId, register: $register, schema: $uitvoeringSchema);
-        if (is_array($uitvoering) === false) {
-            return null;
-        }
-
-        $aanvraagId = (string) ($uitvoering['subsidieaanvraag'] ?? '');
-        if ($aanvraagId === '') {
-            return null;
-        }
-
-        $aanvraag = $objectService->find($aanvraagId, register: $register, schema: $aanvraagSchema);
-        if (is_array($aanvraag) === false) {
-            return null;
-        }
-
-        $caseId = (string) ($aanvraag['case'] ?? '');
-        if ($caseId === '') {
-            return null;
-        }
-
-        return $caseId;
-    }//end resolveLinkedCaseId()
-
-    /**
-     * Whether the decoded kosten list already carries an entry for this
-     * vaststelling (idempotency guard).
-     *
-     * @param array<int, mixed> $kosten         Decoded kosten entries.
-     * @param string            $vaststellingId The settlement id to look for.
-     *
-     * @return bool
-     */
-    private function hasExistingEntry(array $kosten, string $vaststellingId): bool
-    {
-        foreach ($kosten as $entry) {
-            if (is_array($entry) === false) {
-                continue;
-            }
-
-            $source = (string) ($entry['source'] ?? '');
-            $id     = (string) ($entry['vaststellingId'] ?? '');
-            if ($source === self::KOSTEN_SOURCE && $id === $vaststellingId) {
-                return true;
-            }
-        }
-
-        return false;
-    }//end hasExistingEntry()
-
-    /**
-     * Decode a case's raw `kosten` field (array or JSON-encoded string) into
-     * a plain list, defaulting to an empty list for any other shape.
-     * Mirrors {@see \OCA\Procest\Service\Iv3ReportService::decodeKosten()}.
-     *
-     * @param mixed $raw The raw `kosten` field value.
-     *
-     * @return array<int, mixed>
-     */
-    private function decodeKosten(mixed $raw): array
-    {
-        if (is_array($raw) === true) {
-            return $raw;
-        }
-
-        if (is_string($raw) === false || $raw === '') {
-            return [];
-        }
-
-        $decoded = json_decode($raw, true);
-        if (is_array($decoded) === true) {
-            return $decoded;
-        }
-
-        return [];
-    }//end decodeKosten()
 
     /**
      * Resolve the ObjectService and register/schema ids.
